added path.normalize to sourceRoot
There were errors on my windows system caused by the else branch in line 77-83. 
You want to replace the sourceRoot with the output relativePath, I guess it was not found. So all my files ended up in the entry folder and not in the specified output folder.
This plugin works fine on os/linux. 

Thank you for creating this plugin :)